getting the text to Emotes to show on the screen
