+++ app/src/test/java/com/example/clicker/ParsingEngineTest.kt
@@ -1,5 +1,6 @@
 package com.example.clicker
 
+import com.example.clicker.network.websockets.MessageScanner
 import com.example.clicker.network.websockets.MessageType
 import com.example.clicker.network.websockets.ParsingEngine
 import com.example.clicker.network.websockets.findEmoteNames
@@ -266,77 +267,5 @@ class ParsingEngineTest {
 }
 
 
-enum class PrivateMessageType {
-    MESSAGE, EMOTE
-}
-data class MessageToken(
-    val messageType:PrivateMessageType,
-    val messageValue:String ="",
-    val url:String=""
-)
-
-class MessageScanner(
-    private val source:String
-){
-    private val map = hashMapOf<String, MessageToken>()
-
-    private val tokens = mutableListOf<MessageToken>()
-    val tokenList:List<MessageToken> = tokens
-
-    init {
-        map["SeemsGood"] = MessageToken(
-            messageType = PrivateMessageType.EMOTE,
-            url = "https://static-cdn.jtvnw.net/emoticons/v2/64138/static/light/1.0"
-        )
-    }
-
-    private var start = 0
-    private var current =0
-
-     fun startScanningTokens(){
-        while(!isAtEnd()){
-            start = current
-            scanToken()
-        }
-    }
-
-    private fun scanToken(){
-        val c = advance()
-        when(c){
-            ' '->{}
-            else->{
-                while (notEndNullOrEmptySpace(peek())){
-                    advance()
-                }
-                addToken()
-            }
-
-        }
-    }
-
-    private fun addToken(){
-        val text = source.substring(start, current)
-        val tokenType = map[text]?:  MessageToken(messageType = PrivateMessageType.MESSAGE, messageValue = text)
-        tokens.add(tokenType)
-    }
-
-    private fun notEndNullOrEmptySpace(c:Char):Boolean{
-        return !isAtEnd() && c != '\u0000' && c != '\u0020'
-    }
-
-    private fun isAtEnd():Boolean{
-        return current>=source.length
-    }
-    private fun advance(): Char {
-        return source[current++]
-    }
-
-    private fun peek(): Char {
-        return if (isAtEnd()) '\u0000' else source[current]
-    }
-
-
-
-}
